+++ packages/client/src/forms/advancedSearch/fieldDefinitions/Birth.ts
@@ -223,8 +223,7 @@ export const advancedSearchBirthSectionEventDetails: IFormSectionGroup = {
       },
       dynamicOptions: {
         resource: 'locations',
-        dependency: 'eventCountry',
-        initialValue: 'agentDefault'
+        dependency: 'eventCountry'
       },
       conditionals: [
         {
@@ -259,8 +258,7 @@ export const advancedSearchBirthSectionEventDetails: IFormSectionGroup = {
       },
       dynamicOptions: {
         resource: 'locations',
-        dependency: 'eventLocationLevel1',
-        initialValue: 'agentDefault'
+        dependency: 'eventLocationLevel1'
       },
       conditionals: [
         {